Implemented more IDE event listeners, implemented time measuring for other activities ("building" and "debugging")
Enhancements proposed on wakatime/visualstudio-wakatime#106

**Note on .csproj reference:**
I had to remove the original reference so the signed extension utils dll and replace with a reference to my locally built file. I will obviously not gonna commit the .csproj file with the reference change, which means this code **will only run** with the original reference once the pull request wakatime/nuget-wakatime-shared-extension-utils#3 has been merged and the binary rebuilt.

I will further document the changes on the PR.